runtime: fix uint64 atomic fallback return
Why: avoid truncating 64-bit counter values on builds without 64-bit atomic builtins.

Impact: no behavior change on normal atomic builds; no-atomic fallbacks preserve the full returned uint64 value.

Before/After: the mutex fallback returned unsigned; it now returns uint64 like the atomic helper.

Technical Overview:

- Update ATOMIC_INC_AND_FETCH_uint64 fallback signature to return uint64.

- Keep the existing uint64 local value and mutex-protected increment logic unchanged.

- Align the fallback ABI with the HAVE_ATOMIC_BUILTINS64 implementation.

- This is intentionally scoped to the no-atomic 64-bit helper path.

With the help of AI-Agents: Codex

+++ runtime/atomic.h
@@ -375,7 +375,7 @@ static inline void ATOMIC_STORE_time_t_RELAXED(time_t *data, pthread_mutex_t *ph
     pthread_mutex_unlock(phlpmut);
 }
 
-static inline unsigned ATOMIC_INC_AND_FETCH_uint64(uint64 *data, pthread_mutex_t *phlpmut) {
+static inline uint64 ATOMIC_INC_AND_FETCH_uint64(uint64 *data, pthread_mutex_t *phlpmut) {
     uint64 val;
     pthread_mutex_lock(phlpmut);
     val = ++(*data);
